cham/requirements.txt
plotly==5.18.0
streamlit==1.35.0
convertdate==2.4.0
joblib==1.3.2
multiprocess==0.70.16
pandas-gbq==0.19.1
streamlit_extras==0.4.0
numpy
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage